Auto-renames .zip.zip files to get rid of the second .zip When a file cannot be unzipped it now alerts you the zip is invalid. Fixed some corrections not being yellow. Now shows original submit time first when an assignment is modified at a later date. Fixed some crashes involving zip files with no java files Rename will now add a number to the end if you try to rename to a file that already exists File selection is now kept between grading files, it is also kept when sorting the files. If there are no files for you're class period it will now tell you, instead of showing a empty table Fixed renaming files not being synced with sorting files. Table now locks and is un-selectable during certain operations that involve changing data that is in the table in order to eliminate crashes when people click too fast/much Refresh button works better If there are no .java files in the zip it will show you what files are in the zip. Fixed all bugs with deleting files, including a CRITICAL bug where it would delete non-graded assignments (don't use the delete button in versions older than this one) If you somehow do manage to grade a invalid file it wont crash. Optimized startup again, shouldn't do the double refresh thing anymore. Added helpful feedback with deleting files, if all files selected were deleted, if some were, or if none were Moved refresh to background thread, the gui wont freeze when you refresh anymore. Will not put ResubmittedResubmitted after something is resubmitted twice, it will only put it once. Grades now are appeneded with a ` this is to show on the spreadsheet it is a grade that is not in IC.

+++ src/DropboxGrader/WorkerThread.java
@@ -22,11 +22,13 @@ public class WorkerThread implements Runnable{
     private boolean graderAfter;
     private DbxFile fileToRun;
     private int timesToRun;
+    private boolean refreshNeeded;
     public WorkerThread(FileManager man,Gui gui){
         manager=man;
         this.gui=gui;
         fileQueue=new ArrayList();
         deleteQueue=new ArrayList();
+        refreshNeeded=true;
     }
     @Override
     public void run() {
@@ -46,7 +48,8 @@ public class WorkerThread implements Runnable{
                 }
             }
             for(DbxFile f:deleteQueue){
-                f.delete();
+                if(f!=null)
+                    f.delete();
             }
             if(!deleteQueue.isEmpty()){
                 deleteQueue.clear();
@@ -60,6 +63,12 @@ public class WorkerThread implements Runnable{
                 fileToRun.run(gui.getRunner(), timesToRun);
                 fileToRun=null;
                 timesToRun=0;
+            }
+            if(refreshNeeded){
+                manager.refresh();
+                gui.refreshFinished();
+                
+                refreshNeeded=false;
             }
             try {
                 Thread.sleep(100);
@@ -100,6 +109,9 @@ public class WorkerThread implements Runnable{
             return;
         }
         deleteQueue.add(manager.getFile(file));
+    }
+    public void refreshData(){
+        refreshNeeded=true;
     }
     
 }
